% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps.R
\name{mps}
\alias{mps}
\title{Отримайте перелік усіх народних депутатів.}
\usage{
mps(key)
}
\arguments{
\item{key}{Ваш персональний ключ API, отриманий на rada4you.org}
}
\value{
Датафрейм cкладається з чотирьох змінних: mp_id, mp_name, party, electorate
}
\description{
Отримайте перелік усіх народних депутатів.
}
