% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps_policies.R
\name{mps_policies}
\alias{mps_policies}
\title{Отримайте датафрейм із рівнем підтримки певного депутата кожної з політик.}
\usage{
mps_policies(id, key)
}
\arguments{
\item{id}{Вектор, який містить щонайменше один id депутата}

\item{key}{Ваш персональний ключ API, отриманий на rada4you.org}
}
\value{
Датафрейм складається з десяти змінних для кожного депутата: mp_id, mp_name, party, electorate, policy_id, policy_name, description, provisional, agreement, voted
}
\description{
Отримайте датафрейм із рівнем підтримки певного депутата кожної з політик.
}
