% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{policies}
\alias{policies}
\title{Отримайте перелік всіх доступних політик}
\usage{
policies(key)
}
\arguments{
\item{key}{Ваш персональний ключ API, отриманий на rada4you.org}
}
\value{
Датафрейм складається з чотирьох змінних: policy_id, policy_name, description, provisional.
}
\description{
Отримайте перелік всіх доступних політик
}
