% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_info.R
\name{policy_info}
\alias{policy_info}
\title{Отримайте короткий опис політики}
\usage{
policy_info(id = 1, key)
}
\arguments{
\item{id}{Вектор, який містить щонайменше один id політики}

\item{key}{Ваш персональний ключ API, отриманий на rada4you.org}
}
\value{
Датафрейм, що складається з одинадцятьох змінних для кожної політики: policy_id, policy_name, description, provisional, vote_id, vote_name, date, time, aye_votes, no_votes, rebellions.
}
\description{
Отримайте короткий опис політики
}
