% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_mps.R
\name{policy_mps}
\alias{policy_mps}
\title{Отримайте датафрейм із рівнем підтримки політики кожним з нардепів.}
\usage{
policy_mps(id = 1, key)
}
\arguments{
\item{id}{Вектор, який містить містить щонайменше один id політики}

\item{key}{Ваш персональний ключ API, отриманий на rada4you.org}
}
\value{
Датафрейм із рівнем підтримки політики кожним з нардепів, що складається з десяти змінних для кожної політики:  policy_id, policy_name, description, provisional, mp_id, mp_name, party, electorate, agreement, voted.
}
\description{
Отримайте датафрейм із рівнем підтримки політики кожним з нардепів.
}
