% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_info.R
\name{voting_info}
\alias{voting_info}
\title{Отримайте короткий опис голосування}
\usage{
voting_info(id = 4896, key)
}
\arguments{
\item{id}{Вектор, який містить щонайменше один id голосування}

\item{key}{Ваш персональний ключ API, отриманий на rada4you.org}
}
\value{
Датафрейм складається з 14 змінних: vote_id, vote_name, date, time, aye_votes, no_votes, possible_turnout, rebellions, edited, summary, bill_id, bill_number, bill_title, bill_url
}
\description{
Отримайте короткий опис голосування
}
