% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_mps.R
\name{voting_mps}
\alias{voting_mps}
\title{Отримайте датафрейм із голосуванням кожного депутата щодо певного голосування.}
\usage{
voting_mps(id = 4896, key)
}
\arguments{
\item{id}{Вектор, який містить id щонайменше одне id голосування}

\item{key}{Ваш персональний ключ API, отриманий на rada4you.org}
}
\value{
Датафрейм складається з десяти стовбців: vote_id, vote_name, bill_id, bill_number, bill_title, date, mp_name, party, electorate, vote
}
\description{
Отримайте датафрейм із голосуванням кожного депутата щодо певного голосування.
}
