% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_policies.R
\name{voting_policies}
\alias{voting_policies}
\title{Отримайте датафрейм із сукупністю політик до яких підв'язане певне голосування.}
\usage{
voting_policies(id = 4896, key)
}
\arguments{
\item{id}{Вектор, який містить щонайменше один id голосування.}

\item{key}{Ваш персональний ключ API, отриманий на rada4you.org.}
}
\value{
Датафрейм із 10 змінних vote_id, vote_name, date, bill_number,policy_id, policy_name, description, provisional, vote, strong.
}
\description{
Отримайте датафрейм із сукупністю політик до яких підв'язане певне голосування.
}
