% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votings.R
\name{votings}
\alias{votings}
\title{Отримайте датафрейм з усіма голосуваннями за обраний період.}
\usage{
votings(start_date, end_date, key)
}
\arguments{
\item{start_date}{Перший день обраного періоду.}

\item{end_date}{Останній день обраного періоду.}

\item{key}{Ваш персональний ключ API, отриманий на rada4you.org}
}
\value{
Датафрейм, що складається з дев'яти стовбців: vote_id, vote_name, date, time, aye_votes, no_votes, possible_turnout, rebellions, edited
}
\description{
Отримайте датафрейм з усіма голосуваннями за обраний період.
}
